refactor: CQRS 패턴에 따른 LineService 분리
